Update discordexperiments.plugin.js
Based on the comments of [this gist](https://gist.github.com/MeguminSama/2cae24c9e4c335c661fa94e72235d4c4).

The userStore that needs to be patched seems to have moved to a new module. The old module still exists and has similar properties, but it has a property called _version that the new module does not.

I was also getting the error on disable, but I couldn't find a way to fix it. Since it seems to be harmless, I just added a catch for that specific error message.