Fix stop()
I realize now that it throws for the same reason the previous CONNECTION_OPEN does: it expects the user object to have properties relating to experiments. Why this happens on disable as well... I'm not sure.

Thankfully, ExperimentStore has an action handler for logging out, which disables all experiments and doesn't seem to affect anything else.